Update test organization
