+++ orchestration/src/test/java/com/sap/ai/sdk/orchestration/OrchestrationModuleConfigTest.java
@@ -1,27 +1,21 @@
 package com.sap.ai.sdk.orchestration;
 
 import static com.sap.ai.sdk.orchestration.AzureModerationPolicy.ALLOW_SAFE_LOW_MEDIUM;
+import static com.sap.ai.sdk.orchestration.OrchestrationAiModel.GPT_4O;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.sap.ai.sdk.orchestration.client.model.LLMModuleConfig;
+import com.sap.ai.sdk.orchestration.client.model.DPIConfig;
+import com.sap.ai.sdk.orchestration.client.model.DPIEntities;
 import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class OrchestrationModuleConfigTest {
 
-  private OrchestrationModuleConfig config;
-
-  @BeforeEach
-  void setUp() {
-    final var llmModuleConfig =
-        new LLMModuleConfig().modelName("gpt-35-turbo").modelParams(Map.of());
-    config = new OrchestrationModuleConfig().withLlmConfig(llmModuleConfig);
-  }
-
   @Test
   void testStackingInputAndOutputFilter() {
+    final var config = new OrchestrationModuleConfig().withLlmConfig(GPT_4O);
+
     final var filter =
         new AzureContentFilter()
             .hate(ALLOW_SAFE_LOW_MEDIUM)
@@ -40,13 +34,54 @@ class OrchestrationModuleConfigTest {
 
   @Test
   void testThrowOnEmptyFilterConfig() {
-    final var filter = new AzureContentFilter();
 
-    assertThatThrownBy(() -> config.withInputFiltering(filter))
+    final var config = new OrchestrationModuleConfig().withLlmConfig(GPT_4O);
+
+    assertThatThrownBy(() -> config.withInputFiltering(new AzureContentFilter()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage("At least one filter moderation policy must be set");
-    assertThatThrownBy(() -> config.withOutputFiltering(filter))
+    assertThatThrownBy(() -> config.withOutputFiltering(new AzureContentFilter()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage("At least one filter moderation policy must be set");
+  }
+
+  @Test
+  void testDpiMaskingConfig() {
+    var maskingConfig = DpiMasking.anonymization().withEntities(DPIEntities.ADDRESS);
+    var config =
+        new OrchestrationModuleConfig().withLlmConfig(GPT_4O).withMaskingConfig(maskingConfig);
+
+    assertThat(config.getMaskingConfig()).isNotNull();
+    assertThat(config.getMaskingConfig().getMaskingProviders()).hasSize(1);
+    DPIConfig dpiConfig = (DPIConfig) config.getMaskingConfig().getMaskingProviders().get(0);
+    assertThat(dpiConfig.getMethod()).isEqualTo(DPIConfig.MethodEnum.ANONYMIZATION);
+    assertThat(dpiConfig.getEntities()).hasSize(1);
+    assertThat(dpiConfig.getEntities().get(0).getType()).isEqualTo(DPIEntities.ADDRESS);
+
+    var configModified = config.withMaskingConfig(maskingConfig);
+    assertThat(configModified.getMaskingConfig()).isNotNull();
+    assertThat(configModified.getMaskingConfig().getMaskingProviders())
+        .withFailMessage("withMaskingConfig() should overwrite the existing config and not append")
+        .hasSize(1);
+  }
+
+  @Test
+  void testLLMConfig() {
+    Map<String, Object> params = Map.of("foo", "bar");
+    String version = "2024-05-13";
+    OrchestrationAiModel aiModel = GPT_4O.withModelParams(params).withModelVersion(version);
+    var config = new OrchestrationModuleConfig().withLlmConfig(aiModel);
+
+    assertThat(config.getLlmConfig()).isNotNull();
+    assertThat(config.getLlmConfig().getModelName()).isEqualTo(GPT_4O.getModelName());
+    assertThat(config.getLlmConfig().getModelParams()).isEqualTo(params);
+    assertThat(config.getLlmConfig().getModelVersion()).isEqualTo(version);
+
+    assertThat(GPT_4O.getModelParams())
+        .withFailMessage("Static models should be unchanged")
+        .isEmpty();
+    assertThat(GPT_4O.getModelVersion())
+        .withFailMessage("Static models should be unchanged")
+        .isEqualTo("latest");
   }
 }
